requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.2
zipp==3.15.0
python-jenkins>=1.7.0
